Now Camera moves when pressing WASD (not that smooth but usable I'd say)

+++ Source/Core/TempGameData/TempGameData.cpp
@@ -18,7 +18,7 @@ namespace GameData
         CreateCamera();
     }
 
-    void TempGameData::Update(float ms)
+    void TempGameData::Update(double ms)
     {
         camera.UpdateCamera(ms);
     }
